+++ src/lib/square/billing.ts
@@ -6,7 +6,7 @@ import { getValidAccessToken } from './oauth';
 import { centsToMoney } from './money';
 import { platformFeeCents } from './config';
 import { computeNextCharge } from './periods';
-import { recordSuccessfulPayment } from './webhook';
+import { recordSuccessfulPayment } from './payment-recording';
 import type { PaymentKind, RecurringSubscription } from '@prisma/client';
 
 // After this many consecutive failed charges, give up and cancel.
@@ -39,7 +39,8 @@ export function findDueSubscriptions(now: Date, limit = 200): Promise<RecurringS
 export async function chargeSubscriptionNow(subscriptionId: string): Promise<ChargeResult> {
   const sub = await prisma.recurringSubscription.findUnique({ where: { id: subscriptionId } });
   if (!sub) throw new Error('Subscription not found');
-  if (sub.status === 'CANCELLED') return { status: 'CANCELLED', paymentId: null, receiptNumber: null };
+  if (sub.status === 'CANCELLED')
+    return { status: 'CANCELLED', paymentId: null, receiptNumber: null };
 
   let accessToken: string;
   let locationId: string;
@@ -103,7 +104,9 @@ export async function chargeSubscriptionNow(subscriptionId: string): Promise<Cha
     sqPayment = res.payment;
     if (!sqPayment?.id) throw new Error('Square did not return a payment');
   } catch (chargeError) {
-    await prisma.payment.update({ where: { id: payment.id }, data: { status: 'FAILED' } }).catch(() => {});
+    await prisma.payment
+      .update({ where: { id: payment.id }, data: { status: 'FAILED' } })
+      .catch(() => {});
     const attempts = sub.failedAttempts + 1;
     const giveUp = attempts >= MAX_FAILED_ATTEMPTS;
     await prisma.recurringSubscription
@@ -135,23 +138,36 @@ export async function chargeSubscriptionNow(subscriptionId: string): Promise<Cha
   const base = sub.nextChargeAt > new Date() ? sub.nextChargeAt : new Date();
   const next = computeNextCharge(base, sub.interval as 'MONTHLY' | 'ANNUAL');
   try {
-    const applied = await recordSuccessfulPayment({ localPaymentId: payment.id, squarePayment: sqPayment });
+    const applied = await recordSuccessfulPayment({
+      localPaymentId: payment.id,
+      squarePayment: sqPayment,
+    });
     await prisma.recurringSubscription.update({
       where: { id: sub.id },
       data: { status: 'ACTIVE', lastChargedAt: new Date(), nextChargeAt: next, failedAttempts: 0 },
     });
-    return { status: sqPayment.status ?? 'COMPLETED', paymentId: payment.id, receiptNumber: applied.receiptNumber };
+    return {
+      status: sqPayment.status ?? 'COMPLETED',
+      paymentId: payment.id,
+      receiptNumber: applied.receiptNumber,
+    };
   } catch (bookkeepingError) {
     console.error('Subscription charged but local bookkeeping failed; webhook will reconcile', {
       subscriptionId: sub.id,
       paymentId: payment.id,
-      error: bookkeepingError instanceof Error ? bookkeepingError.message : String(bookkeepingError),
+      error:
+        bookkeepingError instanceof Error ? bookkeepingError.message : String(bookkeepingError),
     });
     // Still advance so we never re-charge a card we already charged.
     await prisma.recurringSubscription
       .update({
         where: { id: sub.id },
-        data: { status: 'ACTIVE', lastChargedAt: new Date(), nextChargeAt: next, failedAttempts: 0 },
+        data: {
+          status: 'ACTIVE',
+          lastChargedAt: new Date(),
+          nextChargeAt: next,
+          failedAttempts: 0,
+        },
       })
       .catch(() => {});
     return { status: sqPayment.status ?? 'COMPLETED', paymentId: payment.id, receiptNumber: null };
